:truck: 어플리케이션 시작 함수 분리
